move TestClock tests to ClockSpec
