+++ core-tests/shared/src/test/scala/zio/ClockSpec.scala
@@ -3,6 +3,8 @@ package zio
 import zio.test._
 import zio.test.Assertion._
 
+import java.time.Instant
+
 object ClockSpec extends ZIOBaseSpec {
 
   val someFixedInstant = java.time.Instant.parse("2021-11-13T03:42:09.343835Z")
@@ -37,6 +39,45 @@ object ClockSpec extends ZIOBaseSpec {
       assertZIO(ZIO.collectAll(Chunk(zioEuropeAmsterdamClock.localDateTime, zioAmericaPhoenixClock.localDateTime)))(
         isDistinct
       )
-    }
+    },
+    suite("TestClock")(
+      test("should accept any Instant") {
+        check(Gen.instant) { instant =>
+          for {
+            _   <- TestClock.setInstant(instant)
+            now <- Clock.instant
+          } yield assertTrue(now == instant)
+        }
+      },
+      test("should allow sleep") {
+        for {
+          queue <- Queue.unbounded[Int]
+          fiber <- ZIO.foreachDiscard(1 to 3)(i => ZIO.sleep(1.second) *> queue.offer(i)).fork
+          res   <- ZIO.foreach(Vector.fill(3)(0))(_ => TestClock.adjust(1.second) *> queue.take <*> queue.poll)
+          _     <- fiber.interrupt
+        } yield assertTrue(res == Vector((1, None), (2, None), (3, None)))
+      },
+      test("should allow multiple sleeps") {
+        for {
+          queue <- Queue.unbounded[Int]
+          fiber <- ZIO.foreachDiscard(1 to 3)(i => ZIO.sleep(1.second) *> queue.offer(i)).fork
+          _     <- TestClock.adjust(3.seconds)
+          res   <- ZIO.foreach(Vector.fill(3)(0))(_ => queue.take)
+          _     <- fiber.interrupt
+        } yield assertTrue(res == Vector(1, 2, 3))
+      },
+      test("should allow long sleep") {
+        val maxMillisDuration = Duration.fromMillis(Long.MaxValue)
+
+        for {
+          _       <- TestClock.adjust(maxMillisDuration)
+          _       <- TestClock.adjust(maxMillisDuration)
+          promise <- Promise.make[Nothing, Unit]
+          _       <- (ZIO.sleep(maxMillisDuration) *> promise.succeed(())).fork
+          _       <- TestClock.setInstant(Instant.MAX)
+          _       <- promise.await
+        } yield assertTrue(true)
+      }
+    )
   )
 }
